requirements.txt
speciesnet[server]
fastapi[standard]==0.121.2
google-generativeai>=0.7.0
Pillow>=10.0.0
python-dotenv